requirements.txt
torch==2.5.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
invoke==2.2.0
ruff==0.9.1
pytest==8.3.4
numpy==2.2.1
torchvision==0.20.1
coverage==7.6.10
pytorch-lightning==2.5.0.post0
hydra-core==1.3.2
wandb==0.19.3
timm==1.0.13
python-dotenv==1.0.1